chore: refactor onboarding install step components
